+++ Example/Controllers/Extensions/ViewController+MiniAppPurchaseDelegate.swift
@@ -1,0 +1,8 @@
+import MiniApp
+import UIKit
+
+extension ViewController: MiniAppPurchaseDelegate {
+    func purchaseProduct(withId: String, completionHandler: @escaping (Result<MAProductResponse, MAProductResponseError>) -> Void) {
+        miniAppIAPModule?.initiateProductPayment(productId: withId, completionHandler: completionHandler)
+    }
+}
